Move service registration to async_setup (#315)
* Moved service registrations to async_setup
* Added second code owner (@ocrease)
* Added quality_scale.yaml
* Update host ip address if a new one is discovered
