+++ custom_components/heatmiserneo/__init__.py
@@ -3,12 +3,10 @@
 
 import asyncio
 from dataclasses import dataclass
-from datetime import timedelta
 import logging
 from typing import Any
 
 from neohubapi.neohub import NeoHub, NeoStat
-import voluptuous as vol
 
 from homeassistant.const import CONF_API_TOKEN, CONF_HOST, CONF_PORT, Platform
 from homeassistant.core import CoreState, HomeAssistant
@@ -39,8 +37,13 @@ from .discovery import (
     async_trigger_discovery,
     async_update_entry_from_discovery,
 )
+from .services import async_setup_services
+from .utils import unique_id_is_mac
 
 _LOGGER = logging.getLogger(__name__)
+
+
+CONFIG_SCHEMA = cv.config_entry_only_config_schema(DOMAIN)
 
 PLATFORMS = [
     Platform.BINARY_SENSOR,
@@ -85,13 +88,6 @@ class TimerProfileLevel(ProfileLevel):
     state: bool
 
 
-@dataclass
-class RawTimerProfileLevel(ProfileLevel):
-    """Profile level for on/off states."""
-
-    end_time: str
-
-
 async def async_setup(hass: HomeAssistant, hass_config: ConfigType) -> bool:
     """Set up the Heatmiser Neo integration."""
 
@@ -103,6 +99,9 @@ async def async_setup(hass: HomeAssistant, hass_config: ConfigType) -> bool:
     async_track_time_interval(
         hass, _async_discovery, DISCOVERY_INTERVAL, cancel_on_shutdown=True
     )
+
+    async_setup_services(hass)
+
     return True
 
 
@@ -174,20 +173,6 @@ async def options_update_listener(
 ):
     """Handle options update."""
     await hass.config_entries.async_reload(config_entry.entry_id)
-
-
-def time_period_minutes(value: float | str) -> timedelta:
-    """Validate and transform minutes to a time offset."""
-    try:
-        return timedelta(minutes=float(value))
-    except (ValueError, TypeError) as err:
-        raise vol.Invalid(f"Expected minutes, got {value}") from err
-
-
-hold_duration_validation = vol.All(
-    vol.Any(cv.time_period_str, time_period_minutes, timedelta, cv.time_period_dict),
-    cv.positive_timedelta,
-)
 
 
 async def _async_migrate_unique_ids(
@@ -297,13 +282,6 @@ async def _async_migrate_unique_ids(
             entity_registry.async_update_entity(
                 entity_id=reg_entry.entity_id, new_unique_id=new_unique_id
             )
-
-
-def unique_id_is_mac(unique_id: str | None) -> bool:
-    "Check if a unique id is a mac address."
-    if not unique_id:
-        return False
-    return unique_id.count(":") == 5 and len(unique_id) == 17
 
 
 def _has_old_identifier(device: dr.DeviceEntry) -> bool:
