+++ custom_components/heatmiserneo/quality_scale.yaml
@@ -1,0 +1,118 @@
+rules:
+  # Bronze
+  action-setup: done
+  appropriate-polling:
+    status: done
+    comment: |
+      Polling set to 30 seconds
+  brands:
+    status: done
+    comment: |
+      Assets are under custom_integrations
+  common-modules:
+    status: done
+    comment: |
+      Coordinator, base entity and use of entity descriptors are used extensively
+  config-flow-test-coverage: todo
+  config-flow: done
+  dependency-transparency:
+    status: done
+    comment: |
+      neohubapi is available on PiPy
+  docs-actions:
+    status: todo
+    comment: |
+      No custom actions are defined.
+  docs-high-level-description: todo
+  docs-installation-instructions: todo
+  docs-removal-instructions: todo
+  entity-event-setup:
+    status: done
+    comment: |
+      No explicit event subscriptions.
+  entity-unique-id:
+    status: done
+    comment: |
+      See entity.py
+  has-entity-name:
+    status: done
+    comment: |
+      See entity.py
+  runtime-data:
+    status: done
+    comment: |
+      See coordinator.py
+  test-before-configure:
+    status: done
+    comment: |
+      See coordinator.py
+  test-before-setup: todo
+  unique-config-entry:
+    status: done
+    comment: |
+      Where possible we use the mac address of the hub. If that is not available (mac address is only available with the web socket connection) we check the host IP as fallback. Auto discovered hubs also have the mac address
+  # Silver
+  action-exceptions:
+    status: todo
+    comment: |
+      No custom actions are defined.
+  config-entry-unloading: done
+  docs-configuration-parameters: todo
+  docs-installation-parameters: todo
+  entity-unavailable: done
+  integration-owner: done
+  log-when-unavailable:
+    status: todo
+    comment: |
+      Handled by coordinator.
+  parallel-updates: todo
+  reauthentication-flow:
+    status: todo
+    comment: |
+      No authentication required.
+  test-coverage: todo
+  # Gold
+  devices: done
+  diagnostics: done
+  discovery-update-info: done
+  discovery:
+    status: done
+    comment: |
+      Zeroconf and a proprietary Heatmiser discovery protocol
+  docs-data-update: todo
+  docs-examples: todo
+  docs-known-limitations: todo
+  docs-supported-devices: todo
+  docs-supported-functions: todo
+  docs-troubleshooting: todo
+  docs-use-cases: todo
+  dynamic-devices: done
+  entity-category: done
+  entity-device-class: done
+  entity-disabled-by-default: done
+  entity-translations: todo
+  exception-translations:
+    status: todo
+    comment: |
+      No custom exceptions.
+  icon-translations: done
+  reconfiguration-flow:
+    status: todo
+    comment: |
+      Only parameter that could be changed (MAC = unique_id) would force a new config entry.
+  repair-issues:
+    status: done
+    comment: |
+      Only one repair flow, added to handle an upgrade from earlier versions. Likely not required anymore
+  stale-devices:
+    status: done
+    comment: |
+      Devices removed from the hub are automatically removed. Furthermore, async_remove_config_entry_device is implemented to also remove the device from the hub
+
+  # Platinum
+  async-dependency: done
+  inject-websession:
+    status: exempt
+    comment: |
+      Connections are either via websocket or TCP
+  strict-typing: todo
